fixed serious bug in unitary simulators: noise was not calcluated correctly due to missing floating point conversion
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@4088 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair1/SIMULATION/LTE_PHY/pbchsim.c
@@ -880,7 +880,7 @@ int main(int argc, char **argv) {
 	  }
       }
       
-      sigma2_dB = 10*log10((double)tx_lev) +10*log10(PHY_vars_eNb->lte_frame_parms.ofdm_symbol_size/(12*NB_RB)) - SNR;
+      sigma2_dB = 10*log10((double)tx_lev) +10*log10((double)PHY_vars_eNb->lte_frame_parms.ofdm_symbol_size/(double)(12*NB_RB)) - SNR;
       if (n_frames==1)
 	printf("sigma2_dB %f (SNR %f dB) tx_lev_dB %f,%f,%f\n",sigma2_dB,SNR,
 	       10*log10((double)tx_lev),
